refactor: barallel export added on utils
