+++ src/Utilities/index.js
@@ -1,0 +1,2 @@
+export * from './GlobalStyles/GlobalStyles'
+export {getDesignTokens} from './theme'
